Mehmet Emin Ketriz
Body mass index, also known as height-weight index, is a number calculated using a person's weight and height values. The calculation is made by taking body weight in kilograms and height in meters used in the formula. With the value obtained as a result of the body mass index calculation, a classification is made such as whether the person is underweight, normal weight, slightly overweight or obese. However, body fat ratio is not measured with this value. Only a general picture is drawn when determining the amounts of values such as fat, muscle and water. The main purpose of use is to determine whether a person is at a healthy body weight according to his height. In short, body mass index is a measure that shows whether a person is overweight or underweight; But it does not give us values such as fat and muscle in the body.